fix: refactored mobile-redirect component
