+++ packages/cfd/src/Containers/mt5-mobile-redirect-option.scss
@@ -14,6 +14,7 @@
         padding: 0.8rem;
         border-radius: 0.8rem;
         background-color: var(--general-1);
+        text-decoration: none;
         .center {
             gap: 0.8rem;
             width: 100%;
